edited routes, fixed minor bugs
