+++ index.html
@@ -1,5 +1,5 @@
 <!doctype html>
-<html lang="en" class="overflow-hidden">
+<html lang="en" >
 
 <head>
   <meta charset="UTF-8" />
