+++ src/App.css
@@ -142,6 +142,3 @@
   }
 }
 
-.swiper-pagination {
-  margin-top: 120px !important; /* Adjust this value */
-}
